+++ textdirectory/cli.py
@@ -16,15 +16,16 @@ available_transformations = [transformation for transformation in dir(transforma
 
 @click.command()
 @click.option('--directory', help='The directory containing text files', type=str)
-@click.option('--output_file', help='The file to aggregate to', type=str)
+@click.option('--output_file', help='The file to aggregate to', type=str, default='aggregate.txt')
 @click.option('--filetype', help='The file type to look for.', default='txt', type=str)
+@click.option('--encoding', help='The encoding of the files.', default='utf8', type=str)
 @click.option('--recursive', help='Recursion', type=bool)
 @click.option('--filters', help=f'The filters you want to apply. Filters: {available_filters}', type=str)
 @click.option('--transformations', help=f'The transformations you want to apply. '
                                         f'Tranformations: {available_transformations}', type=str)
-def main(directory, output_file, filetype, recursive, filters, transformations):
+def main(directory, output_file, filetype, encoding, recursive, filters, transformations):
     """Console script for textdirectory."""
-    if not directory or not output_file:
+    if not directory:
         click.echo('Welcome to TextDirectory!\nRun textdirectory --help for further information.')
         sys.exit()
 
@@ -40,8 +41,13 @@ def main(directory, output_file, filetype, recursive, filters, transformations):
         for transformation in transformations:
             transformations_list.append(transformation.split(','))
 
-    td = textdirectory.TextDirectory(directory=directory)
-    td.load_files(recursive, filetype)
+    td = textdirectory.TextDirectory(directory=directory, encoding=encoding)
+    try:
+        td.load_files(recursive=recursive, filetype=filetype)
+    except FileNotFoundError:
+        click.echo('There seem to be no files. Maybe you want to run with --recursive True.')
+        sys.exit()
+
     if filters and len(filters_list) > 0:
         td.run_filters(filters_list)
     if transformations and len(transformations_list) > 0: