+++ HISTORY.rst
@@ -30,5 +30,11 @@ History
 * added transformation_remove_non_alphanumerical
 * added filter_by_similar_documents
 
+0.1.4 (2018-04-02)
+------------------
+* fixed an object mutation problem in the tabulate function
+
 0.2.0 (2018-00-00)
 ------------------
+* added the transform_to_memory() function
+* fixed a bug where load_files() would fail if there were no files